update streamlit app
