--- App/load_data.py
@@ -1,5 +1,0 @@
-import pandas as pd
-
-def load_data(file_path):
-    return pd.read_csv(file_path)
-
